Revert "PR : refactored eth cron"
