+++ Ethereum/index.js
@@ -4,7 +4,7 @@ var app = express();
 var mongoose = require('mongoose');
 var server = require('http').createServer(app);
 var bodyParser = require('body-parser');
-var config = require("./config");
+var config = require("./config/common.js").info;
 
 
 mongoose.Promise = global.Promise;
@@ -12,7 +12,7 @@ mongoose.connect(config.db, function(err, db) {
 	if(err) throw err;
 
 	// listen the server port
-	var port = config.port || 80;
+	var port = process.env.PORT || 80;
 	
 	server.listen(port, function () {
 	  console.log('Server listening at port %d', port);
