+++ Ethereum/services/cronservice.js
@@ -1,5 +1,5 @@
-var config = require("../config");
-var web3 = require('../config').web3;
+var config = require("../config/common.js").info;
+var web3 = require('../config/common').web3;
 
 var TransactionModel = require("../model/transactions");
 var TokenModel = require("../model/tokens");
@@ -7,7 +7,7 @@ var ServiceInofModel = require("../model/serviceinfo");
 
 
 var lastCheckedBlock = 0;
-var lastCheckedIndex = -1;
+var lastCheckedIndex = 0;
 var cronServiceInfo = null;
 
 var fs = require('fs');
@@ -36,18 +36,20 @@ async function getLastCheckedBlock() {
 }
 
 async function saveCronServiceInfo() {
-  try {
-    var info = await ServiceInofModel.findOne({});
-    if (info) {
-      info.set({ lastblock: lastCheckedBlock, lastTxnIndex: lastCheckedIndex });
+  await ServiceInofModel.findOne(async function (e, info) {
+    if (!e) {
+      if (info) {
+        info.set({ lastblock: lastCheckedBlock, lastTxnIndex: lastCheckedIndex });
+      }
+      else {
+        info = new ServiceInofModel({ lastblock: lastCheckedBlock, lastTxnIndex: lastCheckedIndex });
+      }
+      await info.save();
     }
     else {
-      info = new ServiceInofModel({ lastblock: lastCheckedBlock, lastTxnIndex: lastCheckedIndex });
+      filelog('saveCronServiceInfo:error: ', e); // Should dump errors here
     }
-    await info.save();
-  } catch (error) {
-    filelog('saveCronServiceInfo:error: ', e); // Should dump errors here
-  }
+  });
 }
 
 async function CheckUpdatedTransactions() {
@@ -55,50 +57,75 @@ async function CheckUpdatedTransactions() {
     var number = await web3.eth.getBlockNumber();
     if (!number) return;
 
-    try {
-      var blockdata = await web3.eth.getBlock(lastCheckedBlock, true);
-      var txnCount = blockdata.transactions.length;
+    var limit = lastCheckedBlock + config.CRON_TREAT_MAX_BLOCKS;
+    for (let i = lastCheckedBlock; i <= number && i <= limit; i++) {
+      try {
+        var blockdata = await web3.eth.getBlock(i, true);
+        var txnCount = blockdata.transactions.length;
 
-      for (let j = lastCheckedIndex + 1; j < txnCount; j++) {
-        let transaction = blockdata.transactions[j];
-        let hash = transaction.hash;
-        let from = transaction.from;
-        let to = transaction.to;
-        let value = transaction.value;
-        var timestamp = blockdata.timestamp;
+        for (let j = lastCheckedIndex + 1; j < txnCount; j++) {
+          let transaction = blockdata.transactions[j];
+          let hash = transaction.hash;
+          let from = transaction.from;
+          let to = transaction.to;
+          let value = transaction.value;
+          var timestamp = blockdata.timestamp;
 
-        let gasprice = transaction.gasPrice;
-        var txnReceipt = await web3.eth.getTransactionReceipt(hash);
-        let fee = gasprice * txnReceipt.gasUsed;
+          let gasprice = transaction.gasPrice;
 
-        var newTxn = await TransactionModel.findOne({ hash });
-        if (!newTxn) {
+          try {
+            var txnReceipt = await web3.eth.getTransactionReceipt(hash);
+          }
+          catch (e) {
+            filelog('getTransactionReceipt: error: ', e); // Should dump errors here
+            return;
+          }
+
+
+          let fee = gasprice * txnReceipt.gasUsed;
+
           var newTxn = new TransactionModel({
-            blocknumber: lastCheckedBlock,
-            hash,
-            from,
-            to,
-            value,
-            fee,
-            timestamp
+            blocknumber: i,
+            hash: hash,
+            from: from,
+            to: to,
+            value: value,
+            fee: fee,
+            timestamp: timestamp
           });
-          await newTxn.save();
-        }
 
-        // save
-        lastCheckedIndex = j;
+          try {
+            await newTxn.save();
+          }
+          catch (e) {
+            filelog('newTxn.save: error: ', e); // Should dump errors here
+            return;
+          }
+
+
+          if (lastCheckedBlock != i || lastCheckedIndex != j) {
+            // filelog("Updating block: " + i);
+            lastCheckedBlock = i;
+            lastCheckedIndex = j;
+
+            await saveCronServiceInfo();
+          }
+        }
+      }
+      catch (e) {
+        filelog('getBlock: error: ', e); // Should dump errors here
+        return;
+      }
+
+      if (lastCheckedBlock != i || lastCheckedIndex != -1) {
+        // filelog("Updating block: " + i);
+        lastCheckedBlock = i;
+        lastCheckedIndex = -1;
+
         await saveCronServiceInfo();
       }
-    }
-    catch (e) {
-      filelog('getBlock: error: ', e); // Should dump errors here
-      return;
-    }
 
-    // save
-    lastCheckedBlock++;
-    lastCheckedIndex = -1;
-    await saveCronServiceInfo();
+    }
   } catch (error) {
     filelog('getBlockNumber: we have a promblem: ', error); // Should dump errors here
   }
@@ -114,4 +141,5 @@ exports.start_cronService = async function () {
 
   await getLastCheckedBlock();
   transactionService();
+  // setInterval(transactionService, config.CRON_TIME_INTERVAL);
 }
