+++ Ethereum/unittest/testapi.js
@@ -2,7 +2,7 @@ var EthereumController = require('../controllers/ethereum');
 var Web3 = require('web3');
 
 // Show Web3 where it needs to look for a connection to Ethereum.
-var config = require('../config');
+var config = require('../config/common').info;
 var web3 = new Web3(new Web3.providers.HttpProvider(config.provider));
